+++ srcs/executer/manager_execution.c
@@ -6,27 +6,11 @@
 /*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/19 11:37:28 by codespace         #+#    #+#             */
-/*   Updated: 2023/09/23 09:59:46 by codespace        ###   ########.fr       */
+/*   Updated: 2023/09/23 10:18:44 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-
-
-/*
-
-	pipes_forks_and_conditionals
-	Sets up the commands based on the big 3 operators: && || and |.
-		-   Saves the read end of the previous pipe, if a pipe was open
-			the next operator may be a pipe as well: pipe will overwrite
-			the previous file descriptors
-		-   Opens a new pipe if the current command will write too a pipe;
-		-   In case of && or ||, waits for the previous commands to arrive
-			in order to determine their exit status and decide wether
-			the following command will execute or not;
-
-*/
 
 int	pipes_forks_and_conditionals(t_block *manager, int index)
 {
@@ -93,53 +77,22 @@ int	waiting_for_my_workers(t_block *manager, int index)
 {
 	int	i;
 
-	//check_for_signals(manager->ms);
 	i = 0;
-	//printf("manager [%s] waiting for %d children \n", manager->prompt, manager->op_count + 1);
-	//printf("i am [%s], starting status %d, waiting, mypid %d\n", manager->prompt, manager->my_status, getpid());
-	//print_worker_pids(manager);
-	//printf("global here %d\n", g_signal);
 	while (i < index)
 	{
-		
-		//printf("child [%s], index, %d, has pid? %d\n", manager->worker_list[i]->prompt, i, manager->worker_pids[i]);
 		if (manager->worker_pids[i] != 0)
 		{
-			//int status;
-			//ms_prepare_signal(manager->ms, SIG_IGN);
-			//printf("my lvl id (%d, %d), waiting for pid %d, my status now is: %d  ", manager->my_level, manager->my_id, manager->worker_pids[i], manager->my_status);
-			wait_and_save_status(manager->worker_pids[i], &manager->my_status, manager->ms->errfd);
-			/*if (waitpid(manager->worker_pids[i], &status, 0) == -1)
-					perror("waitpid");
-			//dprintf(2, "manager %d received pid %d\n", getpid(), manager->worker_pids[i]);
-			if (WIFEXITED(manager->my_status))
-				manager->my_status = WEXITSTATUS(manager->my_status);
-			else if (WIFSIGNALED(status))
-			{
-				if (WTERMSIG(manager->my_status) == SIGINT)
-				{
-					//manager->hit_sigint = 1;
-					ft_putstr_fd("\n", manager->ms->errfd);
-				}
-				manager->my_status = WTERMSIG(manager->my_status) + EXIT_SIGNALED;
-				
-			}*/
-			//printf("global is %d\n", g_signal);	
-			//printf("  and changed to %d i received from child (%d, %d)\n", manager->my_status, manager->my_level +1, i);
+			wait_and_save_status(manager->worker_pids[i], \
+								&manager->my_status, manager->ms->errfd);
 			manager->worker_pids[i] = 0;
-			//ms_prepare_signal(manager->ms, signal_handler);
 		}
 		else if (manager->worker_exit_status[i] >= 0)
 		{
-			//printf("my lvl id (%d, %d) original status %d will change to %d from child wnumber (%d,%d)\n", manager->my_level, manager->my_id, manager->my_status, manager->worker_exit_status[i], manager->my_level +1, i);
 			manager->my_status = manager->worker_exit_status[i];
 			manager->worker_exit_status[i] = -1;
 		}
-		//else
-		//	printf("no pid, no manual status\n");
 		i++;
 	}
-	//printf("ending status %d, moving on, mypid %d\n", manager->my_status, getpid());
 	if (!manager->i_am_forked && manager->my_manager)
 		manager->my_manager->my_status = manager->my_status;
 	manager->ms->exit_status = manager->my_status;
